Implemented swipe mechanic for right arm
Added a swipe mechanic for the right arm that functions by pressing and holding a button on the console belonging to the mech. A lot of work was done on the BP_ProtoMech and BP_ThePlayerCharacter for pressing buttons. Need to add functionality for the left arm, and work on the jab mechanic