Added Test Enemies, Swipe Collision, Spine Progression and Enemy Spawning
Added placeholder enemies that spawn in and move towards the player, the function that determines their spawn location needs work. Added collision to the arms, currently the enemies seem to be colliding even without the enemy attacking which needs to be fixed. Added progression to the movement of the mech along the spline that stops at certain points, spawns enemies, then when all enemies are defeated resumes.