Brought button UI prompts to the current branch [WARNING - READ COMMIT]
[Warning below]
There were a few errors compiling things that I had to disconnect to test my work, tried to reconnect them after I was done but I'm unsure if I got them all. May be worth double checking the protomech, player character and also the left and right arms to see if things are okay.

Added UI prompts to pressing buttons in the mech with box colliders. This may not be an efficient way to do this so let me know if it needs changing.